Fix typing errors and unused imports in split expression mixins
- Moved `_guess_type` and `_infer_generator_types` directly into `TranslatorBase` to ensure all child mixins pass `mypy` static typing without referencing undefined attributes.
- Added type annotations for remaining shared properties (`used_string_format`, `global_vars`, `dataclasses`) to `TranslatorBase.__init__`.
- Cleaned up unused typing and custom imports in `expressions_split` child mixin files.
