+++ py2v_transpiler/core/translator/expressions_split/attributes.py
@@ -1,7 +1,5 @@
 import ast
-from typing import List, Optional, Any
 from ..base import TranslatorBase
-from py2v_transpiler.models.v_types import map_python_type_to_v
 
 class AttributesMixin(TranslatorBase):
     def visit_Attribute(self, node: ast.Attribute) -> str:
