+++ py2v_transpiler/core/translator/expressions_split/calls.py
@@ -1,5 +1,5 @@
 import ast
-from typing import List, Optional, Any
+from typing import List, Any
 from ..base import TranslatorBase
 from py2v_transpiler.models.v_types import map_python_type_to_v
