+++ py2v_transpiler/core/translator/expressions_split/subscripts.py
@@ -1,7 +1,6 @@
 import ast
-from typing import List, Optional, Any
+from typing import Any
 from ..base import TranslatorBase
-from py2v_transpiler.models.v_types import map_python_type_to_v
 
 class SubscriptsMixin(TranslatorBase):
     def visit_Subscript(self, node: ast.Subscript) -> str:
